+++ configuration/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+- add handling for default keys `TRANSACTIONSIGNERS_DEFAULT_{KEY,ID,REGION}` and `RPCS_DEFAULT_RPCSTYLE`
+- add tests for new default config keys
+
 ### v0.1.0-rc.23
 
 - fix typo in staging goerli rpc url